Disable meow again
When I used it, it wasn't adding to much of a benefit over using more suitable (faster) emacs bindings.
I think the main gain is as usual, learning the things that exist well instead of venturing away for so long.

For example instead of C-f use the word sexp or paragraph versions.
Instead of C-a which is bad for the pinky, use M-m which is usually what I want anyway.